Add .orig to gitignore, use weights to get the layout proper
I added .orig files to the gitignore as these are files which are
created as backup during merge conflict resolution by kdiff3.

I have succesfully understood a part of how to use context and
get them inside various classes. I used Login's context inside
the EmployeeSender class to lauch a intent as well as to display
a toast.

I used weights to get the various UI elements to be displayed in
proportion to the screen size.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>