Add location retrieval class
I have finished client side coding up to the part where we finalize
the location of the volunteer. The string which the user enter's
is taken and send to the server, where a string consisting of
possible addresses are send using '#' as a delimiter.
we display this in a list.

We need to imporove the entire look of the app

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>